set_vc_size command
