+++ crack-core/src/db/mod.rs
@@ -1,9 +1,11 @@
 pub mod guild;
+pub mod metadata;
 pub mod play_log;
 pub mod playlist;
 pub mod user;
 
 pub use guild::*;
+pub use metadata::*;
 pub use play_log::*;
 pub use playlist::*;
 pub use user::*;
